Backout migration of events on ScrollView to be bridgeless
Summary:
This diff back out the migration of events on ScrollView to be bridgeless (D18862857)
This is causing the bug: T59882485 and the migration is not necessary in production yet.

Heads-up ejanzer this diff will affect events for scroll in Venice, I will create another diff to figure it out the root cause. But we need the land and pick this diff in v252

Changelog: [internal]

Reviewed By: ejanzer

Differential Revision: D19269906

fbshipit-source-id: d61e67975d508cddafc3af34b0f7dfe83bda6bd7