Add links to issue template (#28079)
Summary:
I just stumbled upon [This PR](https://github.com/facebook/react/pull/18039) and thought it would be great for react-native repo to have the same.

This creates a link in issue template that will lead people who found error in the documentation to the appropriate repo, and people who has questions about react-native will be lead to the support page.

<img width="1034" alt="Screen Shot 2020-02-15 at 13 44 02" src="https://user-images.githubusercontent.com/6936373/74582027-ce9bf600-4ff9-11ea-9dc5-21601c4007c4.png">

## Changelog

[Internal] [Changed] - Add links to Issue template
Pull Request resolved: https://github.com/facebook/react-native/pull/28079

Test Plan: This pr does not contain any change in the functionality of react-native.

Differential Revision: D19954747

Pulled By: hramos

fbshipit-source-id: 7cce094e2519c0f508858a034f053d058d801f35