- Bump CLI to ^4.1.x (#28172)
Summary:
Upgrading CLI to latest. alloy should have done a local commit on the 0.62 branch and this pr is just meant to sync the status between master and 0.62 branch.

cc grabbou alloy

nit: it seems that the new 4.1.1 CLI version pulls in quite a bit of dependencies, I hope it's ok. It seems it's mostly because of `pretty-format`.

## Changelog

[Internal] [Changed] - Bump CLI to ^4.1.x
Pull Request resolved: https://github.com/facebook/react-native/pull/28172

Test Plan: None

Reviewed By: hramos

Differential Revision: D20097587

Pulled By: cpojer

fbshipit-source-id: ec95498ced374d1e4e6b696f2167fedf66395ac9